+++ source/detect_bars/changelog.md
@@ -1,4 +1,7 @@
 
+**<span style="color:#56adda">0.0.4</span>**
+- add option to set to False adding file to queue so no library plugins can add the file to the queue if it does not have black bars
+
 **<span style="color:#56adda">0.0.3</span>**
 - extract element 0 from width & height